refactor(partners): hoist cross-domain partner-workspace-folder nav to modules/shared
