+++ packages/twenty-apps/internal/twenty-partners/src/modules/opportunity/navigation-menu-items/my-deals.navigation-menu-item.ts
@@ -2,7 +2,7 @@ import { NavigationMenuItemType, defineNavigationMenuItem } from 'twenty-sdk/def
 
 import { MY_DEALS_VIEW_UNIVERSAL_IDENTIFIER } from 'src/modules/opportunity/views/my-deals.view';
 
-import { PARTNER_WORKSPACE_FOLDER_UNIVERSAL_IDENTIFIER } from 'src/modules/partner/directory/navigation-menu-items/partner-workspace-folder.navigation-menu-item';
+import { PARTNER_WORKSPACE_FOLDER_UNIVERSAL_IDENTIFIER } from 'src/modules/shared/navigation-menu-items/partner-workspace-folder.navigation-menu-item';
 
 export default defineNavigationMenuItem({
   universalIdentifier: '783d4920-6bf6-4dbf-b705-228ffcc9a7d7',
