+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/self-service/navigation-menu-items/my-profile.navigation-menu-item.ts
@@ -5,7 +5,7 @@ import {
   MY_PROFILE_PAGE_LAYOUT_ID,
 } from 'src/modules/partner/self-service/constants/my-profile.constants';
 
-import { PARTNER_WORKSPACE_FOLDER_UNIVERSAL_IDENTIFIER } from 'src/modules/partner/directory/navigation-menu-items/partner-workspace-folder.navigation-menu-item';
+import { PARTNER_WORKSPACE_FOLDER_UNIVERSAL_IDENTIFIER } from 'src/modules/shared/navigation-menu-items/partner-workspace-folder.navigation-menu-item';
 
 export default defineNavigationMenuItem({
   universalIdentifier: MY_PROFILE_NAV_ITEM_ID,
