+++ packages/twenty-apps/internal/twenty-partners/src/modules/shared/navigation-menu-items/partner-workspace-folder.navigation-menu-item.ts
